Added retention of literal phrase > N
The lemmatized phrase is great to ensure that similar conecepts are
grouped together. On the other hand, matching it back to what is
in the CDE field becomes complicated.
This fix record the verbatim phrase and combines it back to the
lemmatized phrase.
Code works and the result for parsing the CDEs is uploaded in the
data_files/analysis directory.