Renamed html action to strip, updated html code
Added some options that tell html stripping code to assume that first line of table is header, or not.
Changed the Property class to allow either str or dict in value key (field). This permits expansion of that field with a json "table".
Refactored some of the other code related to html stripping.

At this point the strip/html action behaves nicely

ToDo:
1. Simple python to prepend "x" to underscore key names.
2. Reverse "x_..." to just "_..."
3. Refactor all verbose and verbosity to accommodate logging.debug which is cleaner
4. Add logic to exclude tinyId from file list (allows exclusion of Cd's that have no permissible values)
5. Develop QC logic for finding Cd's that violate some rules (e.g., must have meaningful permissible values):