Pseudo fasta file output
Code is "functional". Generates a pseudo FASTA format file.
Additional tweak to format is that data lines are prefixed by '-' to
ensure that base85 encoding does not have '>' at the beginning of the
data line.

Code needs clean up throughout, but it produces the desired output and
works on all the data.
Repuroposed much of `extract_embed` logic. Pulled out a few routines
as separate functions to allow for better code resuability.

ToDo:
1. Must implement the unicode cleanup. Have notes elsewhere.